coredump: Include file mappings
1. Record path for file-backed map
2. Apply the following Fujitsu's fixes
   * POSTK_TEMP_FIX_59
   * POSTK_TEMP_FIX_61
   * POSTK_TEMP_FIX_62
   * POSTK_TEMP_FIX_63

TODO: Dump mappings of ld-*.so and executable as file maps although
they are not mapped in McKernel.

+++ executer/kernel/mcctrl/mcctrl.h
@@ -116,7 +116,7 @@ struct coretable {
 };
 #else /* POSTK_DEBUG_TEMP_FIX_61 */
 struct coretable {
-	int len;
+	loff_t len;
 	unsigned long addr;
 };
 #endif /* POSTK_DEBUG_TEMP_FIX_61 */
